Module google_artifactregistry1_beta1::api

Structs

ArtifactRegistry
Central instance to access all ArtifactRegistry related resource activities

Binding
Associates members with a role.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

Expr
Represents a textual expression in the Common Expression Language (CEL) syntax. CEL is a C-like expression language. The syntax and semantics of CEL are documented at https://github.com/google/cel-spec. Example (Comparison): title: "Summary size limit" description: "Determines if a summary is less than 100 chars" expression: "document.summary.size() < 100" Example (Equality): title: "Requestor is owner" description: "Determines if requestor is the document owner" expression: "document.owner == request.auth.claims.email" Example (Logic): title: "Public documents" description: "Determine whether the document should be publicly visible" expression: "document.type != 'private' && document.type != 'internal'" Example (Data Manipulation): title: "Notification string" description: "Create a notification string with a timestamp." expression: "'New message received at ' + string(document.create_time)" The exact variables and functions that may be referenced within an expression are determined by the service that evaluates it. See the service documentation for additional information.

File
Files store content that is potentially associated with Packages or Versions.

Hash
A hash of file content.

ListFilesResponse
The response from listing files.

ListLocationsResponse
The response message for Locations.ListLocations.

ListOperationsResponse
The response message for Operations.ListOperations.

ListPackagesResponse
The response from listing packages.

ListRepositoriesResponse
The response from listing repositories.

ListTagsResponse
The response from listing tags.

ListVersionsResponse
The response from listing versions.

Location
A resource that represents Google Cloud Platform location.

Operation
This resource represents a long-running operation that is the result of a network API call.

Package
Packages are named collections of versions.

Policy
An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources. A Policy is a collection of bindings. A binding binds one or more members to a single role. Members can be user accounts, service accounts, Google groups, and domains (such as G Suite). A role is a named list of permissions; each role can be an IAM predefined role or a user-created custom role. For some types of Google Cloud resources, a binding can also specify a condition, which is a logical expression that allows access to a resource only if the expression evaluates to true. A condition can add constraints based on attributes of the request, the resource, or both. To learn which resources support conditions in their IAM policies, see the IAM documentation. JSON example: { "bindings": [ { "role": "roles/resourcemanager.organizationAdmin", "members": [ "user:mike@example.com", "group:admins@example.com", "domain:google.com", "serviceAccount:my-project-id@appspot.gserviceaccount.com" ] }, { "role": "roles/resourcemanager.organizationViewer", "members": [ "user:eve@example.com" ], "condition": { "title": "expirable access", "description": "Does not grant access after Sep 2020", "expression": "request.time < timestamp('2020-10-01T00:00:00.000Z')", } } ], "etag": "BwWWja0YfJA=", "version": 3 } YAML example: bindings: - members: - user:mike@example.com - group:admins@example.com - domain:google.com - serviceAccount:my-project-id@appspot.gserviceaccount.com role: roles/resourcemanager.organizationAdmin - members: - user:eve@example.com role: roles/resourcemanager.organizationViewer condition: title: expirable access description: Does not grant access after Sep 2020 expression: request.time < timestamp('2020-10-01T00:00:00.000Z') - etag: BwWWja0YfJA= - version: 3 For a description of IAM and its features, see the IAM documentation.

ProjectLocationGetCall
Gets information about a location.

ProjectLocationListCall
Lists information about the supported locations for this service.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectLocationRepositoryCreateCall
Creates a repository. The returned Operation will finish once the repository has been created. Its response will be the created Repository.

ProjectLocationRepositoryDeleteCall
Deletes a repository and all of its contents. The returned Operation will finish once the repository has been deleted. It will not have any Operation metadata and will return a google.protobuf.Empty response.

ProjectLocationRepositoryFileGetCall
Gets a file.

ProjectLocationRepositoryFileListCall
Lists files.

ProjectLocationRepositoryGetCall
Gets a repository.

ProjectLocationRepositoryGetIamPolicyCall
Gets the IAM policy for a given resource.

ProjectLocationRepositoryListCall
Lists repositories.

ProjectLocationRepositoryPackageDeleteCall
Deletes a package and all of its versions and tags. The returned operation will complete once the package has been deleted.

ProjectLocationRepositoryPackageGetCall
Gets a package.

ProjectLocationRepositoryPackageListCall
Lists packages.

ProjectLocationRepositoryPackageTagCreateCall
Creates a tag.

ProjectLocationRepositoryPackageTagDeleteCall
Deletes a tag.

ProjectLocationRepositoryPackageTagGetCall
Gets a tag.

ProjectLocationRepositoryPackageTagListCall
Lists tags.

ProjectLocationRepositoryPackageTagPatchCall
Updates a tag.

ProjectLocationRepositoryPackageVersionDeleteCall
Deletes a version and all of its content. The returned operation will complete once the version has been deleted.

ProjectLocationRepositoryPackageVersionGetCall
Gets a version

ProjectLocationRepositoryPackageVersionListCall
Lists versions.

ProjectLocationRepositoryPatchCall
Updates a repository.

ProjectLocationRepositorySetIamPolicyCall
Updates the IAM policy for a given resource.

ProjectLocationRepositoryTestIamPermissionCall
Tests if the caller has a list of permissions on a resource.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the ArtifactRegistry hub.

Repository
A Repository for storing artifacts with a specific format.

SetIamPolicyRequest
Request message for SetIamPolicy method.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

Tag
Tags point to a version and represent an alternative name that can be used to access the version.

TestIamPermissionsRequest
Request message for TestIamPermissions method.

TestIamPermissionsResponse
Response message for TestIamPermissions method.

Version
The body of a version resource. A version resource represents a collection of components, such as files and other data. This may correspond to a version in many package management schemes.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.